+++ packages/mock-server/docker/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @scalar/mock-server-docker
 
+## 0.2.24
+
 ## 0.2.23
 
 ## 0.2.22