+++ packages/openapi-to-markdown/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @scalar/openapi-to-markdown
 
+## 0.5.8
+
 ## 0.5.7
 
 ## 0.5.6